requirements.txt
fastapi==0.110.3
moviepy==1.0.3
numpy==1.26.4
openai==1.24.0
opencv_python_headless==4.9.0.80
protobuf==5.26.1
pydantic==2.7.1
pydub==0.25.1
python-dotenv==1.0.1
pyttsx3==2.90
Requests==2.31.0
streamlit==1.32.2
uvicorn==0.29.0
google-generativeai==0.4.1